Fix dependencies issues
For some unknown reason, the 'grpc' extra dependency of google-api-core
is now both needed at runtime and not installed if not explicitely mentionned
in the dependencies list.

This commit is also bumping the version of some Google libs.

As Google libs expects recent versions of setuptools, we ask to get a newer one.

To conclude, the singer python SDK only works with a specific version of
pytz so we have to explicit it to avoid having pip resolve a version that is
too new.